Refactoring
Refactored much of the code to better support  portfolio information:

Added colored legend for different types of items - directories links and normal information

Gathered about and contact information. May clean this up later as it could most likely be less hard coded and reusable, but for now the functionality is there for all aspects of my portfolio with the exception of my blog which I will be adding soon.

Updated the help section to help users identify different items and directories.

Fixed error that would occur when trying to cd into a non-directory item.

Added inline-text class to better assist in styling certain output.

Set directoryName parameter to false on getInnerHtmlByClassNameAsync() Function.

Moved contact directory to root of the application.

Changed location of Portfolio DOM reference.